requirements.txt
streamlit
scikit-learn
pandas
numpy
nltk
pdfplumber
python-docx